+++ Common/Advanced/EarlyAccess/IRoslynGuard.cs
@@ -1,0 +1,7 @@
+using Microsoft.CodeAnalysis;
+
+namespace MirrorSharp.Advanced.EarlyAccess {
+    internal interface IRoslynGuard {
+        void ValidateCompilation(Compilation compilation, IRoslynSession session);
+    }
+}