+++ Common/Internal/ILanguageSessionExtensions.cs
@@ -1,0 +1,8 @@
+using System;
+using MirrorSharp.Advanced.EarlyAccess;
+
+namespace MirrorSharp.Internal {
+    internal interface ILanguageSessionExtensions {
+        IRoslynGuard? RoslynGuard { get; }
+    }
+}