+++ VisualBasic/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [2.1.1] - 2020-12-17
+
+### Changed
+- Updated to support MirrorSharp.Common 2.2.2
+
 ## [2.1.0] - 2020-09-05
 ## [2.1.0-preview-01] - 2020-08-28